chore: Seperate MapView component
